Added route protection

+++ controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
     console.log(req.body);
     const {name, email, phone} = req.body;
@@ -24,6 +24,7 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id,
     });
 
     res.status(201).json(contact);
@@ -31,7 +32,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
     
     const contact = await Contact.findById(req.params.id);
@@ -45,13 +46,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User does not have permission to update contacts of other users.")
     }
 
 const updatedContact = await Contact.findByIdAndUpdate(
@@ -65,7 +71,7 @@ const updatedContact = await Contact.findByIdAndUpdate(
 
 //@desc Delete contact
 //@route DELETE /api/contacts
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     
@@ -73,6 +79,12 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User does not have permission to delete contacts of other users.")
+    }
+
     await Contact.deleteOne();
     res.status(200).json(contact);
 })
